+++ CHANGELOG.md
@@ -41,6 +41,8 @@ Group changes to describe their impact on the project, as follows:
     - Parameter `module::PushNotification/windowsphone-package-sid` (deprecated in 2.3.0)
     - Parameter `module::PushNotification/windowsphone-application-secret` (deprecated in 2.3.0)
     - Silk audio encoder/decoder is not supported anymore.
+- **Conference:**
+    - Parameter `conference-server/conference-factory-uri` (deprecated in 2.1.0)
 - **EventLogs:**
     - Parameter `event-logs/dir` (deprecated in 2.0.0)
     - Parameter `event-logs/flexiapi-token` (deprecated in 2.3.3)